Download file performance improvement
When downloading the signed file, the full file is no longer downloaded in a byte array on the server before being returned. Instead a stream is returned directly. It is a custom file stream, so that it can clean up the directory after it is done.

If an error occurs the clean up, then that should be handled by the middleware and shouldn't cause an error response, so the catch/logger isn't needed directly.